--- ClusterTests/ClusterTests.swift
@@ -1,21 +1,0 @@
-//
-//  ClusterTests.swift
-//  ClusterTests
-//
-//  Created by Nikita Belosludtcev on 23/05/2018.
-//  Copyright © 2018 efremidze. All rights reserved.
-//
-
-import XCTest
-import Cluster
-
-class ClusterTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-    }
-    
-    override func tearDown() {
-        super.tearDown()
-    }
-}
